# Sawagger/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/invite/create_batch": {
      "post": {
        "description": "批量生成指定数量的邀请码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "批量生成邀请码",
        "parameters": [
          {
            "description": "生成数量",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InviteController.GenerateCodesRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GenerateAndInsertCodesSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GenerateAndInsertCodesErrorResponse"
            }
          }
        }
      }
    },
    "/invite/create_one": {
      "post": {
        "description": "生成一个新的邀请码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "生成单个邀请码",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GenerateAndInsertCodeSuccessResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GenerateAndInsertCodeErrorResponse"
            }
          }
        }
      }
    },
    "/invite/delete": {
      "delete": {
        "description": "根据 ID 列表删除邀请码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "删除邀请码",
        "parameters": [
          {
            "description": "ID 列表",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InviteCodeService.CreateInviteFilter"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.DeleteInviteCodeSuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.DeleteInviteCodeErrorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/invite/get": {
      "get": {
        "description": "根据 ObjectID 或 code 字段获取邀请码信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "获取邀请码",
        "parameters": [
          {
            "type": "string",
            "description": "邀请码",
            "name": "code",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回邀请码信息",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GetInviteCodeErrorResponse"
            }
          },
          "400": {
            "description": "查询失败或未找到",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.GetInviteCodeSuccessResponse"
            }
          },
          "422": {
            "description": "参数错误",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/invite/list": {
      "post": {
        "description": "根据分页参数和状态获取邀请码列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "分页获取邀请码列表",
        "parameters": [
          {
            "description": "分页参数和过滤条件",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InviteCodeService.CreateInviteFilter"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.ListInviteCodesSuccessResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.ListInviteCodesErrorResponse"
            }
          }
        }
      }
    },
    "/invite/reback": {
      "post": {
        "description": "根据 ID 查询邀请码信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "批量回查邀请码",
        "parameters": [
          {
            "description": "邀请码 ID 列表",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InviteController.InviteCodesRequest"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.SuccessResponseDoc"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/invite/update": {
      "put": {
        "description": "更新指定邀请码的状态",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "邀请码管理"
        ],
        "summary": "更新邀请码状态",
        "parameters": [
          {
            "description": "更新数据",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InviteCodeService.UpdateInviteCode"
            }
          },
          {
            "type": "string",
            "description": "Bearer token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.UpdateInviteCodeSuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/inviteCodeSwaggerResponse.UpdateInviteCodeErrorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.ErrorResponseDoc"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "InviteCodeService.CreateInviteFilter": {
      "type": "object",
      "properties": {
        "beginTime": {
          "type": "string"
        },
        "code": {
          "description": "邀请码",
          "type": "string"
        },
        "deleted": {
          "description": "是否删除 1 未删除 2 已删除",
          "type": "integer"
        },
        "endTime": {
          "type": "string"
        },
        "expired": {
          "description": "是否过期 1 未过期 2 已过期",
          "type": "integer"
        },
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "status": {
          "description": "使用状态 1 已使用 2 未过期",
          "type": "integer"
        },
        "username": {
          "description": "绑定的用户名，\"\"为未绑定",
          "type": "string"
        }
      }
    },
    "InviteCodeService.UpdateInviteCode": {
      "type": "object",
      "properties": {
        "days": {
          "description": "过期延后天数",
          "type": "integer"
        },
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "description": "使用状态 1 已使用 2 未过期",
          "type": "integer"
        }
      }
    },
    "InviteController.GenerateCodesRequest": {
      "type": "object",
      "required": [
        "num"
      ],
      "properties": {
        "num": {
          "type": "integer"
        }
      }
    },
    "InviteController.InviteCodesRequest": {
      "type": "object",
      "required": [
        "codeIds"
      ],
      "properties": {
        "codeIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "inviteCodeSwaggerResponse.DeleteInviteCodeErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.DeleteInviteCodeSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "integer",
          "example": 10
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "inviteCodeSwaggerResponse.GenerateAndInsertCodeErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": -1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "null"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.GenerateAndInsertCodeSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": 1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "id string"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "ok"
        }
      }
    },
    "inviteCodeSwaggerResponse.GenerateAndInsertCodesErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.GenerateAndInsertCodesSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "integer",
          "example": 10
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "inviteCodeSwaggerResponse.GetInviteCodeErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.GetInviteCodeSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "{id: codeID, code: code, username: username, status: 1, expired: 1, expiredTime: 2025-01-01 00:00:00, deleted: 1}"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "inviteCodeSwaggerResponse.ListInviteCodesErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.ListInviteCodesSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "{page:1, pagesize: 10, totalPages: 100,data: [id1_Detail, id2_Detail, id3_Detail]}"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "inviteCodeSwaggerResponse.UpdateInviteCodeErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "error msg"
        }
      }
    },
    "inviteCodeSwaggerResponse.UpdateInviteCodeSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "integer",
          "example": 10
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "swaggerResponse.ErrorResponseDoc": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": -1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "null"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "参数错误"
        }
      }
    },
    "swaggerResponse.SuccessResponseDoc": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": 1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "any"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "ok"
        }
      }
    },
...
  }
}